+++ Simbody/src/MobilizedBodyImpl.h
@@ -266,21 +266,21 @@ public:
     }
 
     const Transform& getInboardFrame (const State& s) const {
-        // TODO: these should come from the state if the mobilizer has variable frames
         const SBInstanceVars& iv = getMyMatterSubsystemRep().getInstanceVars(s);
-        return getMyRigidBodyNode().getX_PF();
+        return getMyRigidBodyNode().getX_PF(iv);
     }
     const Transform& getOutboardFrame(const State& s) const {
-        // TODO: these should come from the state if the mobilizer has variable frames
         const SBInstanceVars& iv = getMyMatterSubsystemRep().getInstanceVars(s);
-        return getMyRigidBodyNode().getX_BM();
+        return getMyRigidBodyNode().getX_BM(iv);
     }
 
     void setInboardFrame (State& s, const Transform& X_PF) const {
-        assert(!"setInboardFrame(s) not implemented yet");
+        SBInstanceVars& iv = getMyMatterSubsystemRep().updInstanceVars(s);
+        getMyRigidBodyNode().updX_PF(iv) = X_PF;
     }
     void setOutboardFrame(State& s, const Transform& X_BM) const {
-        assert(!"setOutboardFrame(s) not implemented yet");
+        SBInstanceVars& iv = getMyMatterSubsystemRep().updInstanceVars(s);
+        getMyRigidBodyNode().updX_BM(iv) = X_BM;
     }
 
     const Transform& getBodyTransform(const State& s) const {